+++ src/ime_controller.h
@@ -1,0 +1,7 @@
+#pragma once
+
+class ImeController {
+ public:
+  bool SetOpenStatus(bool open) const;
+};
+