+++ src/resource.h
@@ -1,0 +1,6 @@
+#pragma once
+
+#define IDI_APP_ICON 101
+
+#define ID_TRAY_EXIT 1001
+#define ID_TRAY_START_AT_SIGN_IN 1002